Refactor test utilities into separate module
